Refactor dataset.py to seperate DatasetStorage
